pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"


[tool.poetry]
name = "autoradarr"
description = "Auto scan IMDB, Kinopoisk, etc for new films and add into radarr database"
version = "0.1.0"
license = "MIT"

authors = []

readme = "README.md"

repository = "https://github.com/miheevv/autoradarr"

keywords = []

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = ">=3.6,<3.10"
beautifulsoup4 = "^4.9.3"
pymongo = "^3.11.3"

[tool.poetry.dev-dependencies]
mypy = "^0.800"

flake8-pytest-style = "^1.3"
nitpick = "^0.23"

safety = "^1.10"

pytest = "^6.2"
pytest-cov = "^2.11"
pytest-randomly = "^3.5"
pymongo = "^3.11.3"
pytest-mock = "^3.5.1"
requests-mock = "^1.8.0"
pylint = "^2.7.4"
mongomock = "^3.22.1"
